backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.1",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "openai>=1.0",
    "pypdf>=6.0.0",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "socksio>=1.0.0",
    "uvicorn>=0.38.0",
    # 数据库相关
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "alembic>=1.13.0",
]
